requirements.txt
line-bot-sdk
flask
requests
pandas
datetime
gunicorn==20.1.0
pytz==2022.1
schedule==0.6.0
numpy